Setup for deployment
Update config

debug

Add postbuild script

test

test

test

Update env var

Cleanup and update README

Remove Procfile

Update node command
