+++ .env.example
@@ -1,0 +1,16 @@
+SESSION_SECRET = <input>
+IMGUR_CLIENT_ID = <input>
+IMGUR_CLIENT_SECRET = <input>
+
+TYPEORM_CONNECTION = postgres
+TYPEORM_HOST = <input>
+TYPEORM_PORT = <input>
+TYPEORM_USERNAME = <input>
+TYPEORM_PASSWORD = <input>
+TYPEORM_DATABASE = <input>
+TYPEORM_LOGGING = true
+TYPEORM_ENTITIES = build/entities/*.js
+TYPEORM_MIGRATIONS = build/migrations/*.js
+TYPEORM_SEEDING_SEEDS = build/seeds/**/*{.ts,.js}
+TYPEORM_SEEDING_FACTORIES = build/factories/**/*{.ts,.js}
+TYPEORM_MIGRATIONS_DIR = server/migrations
