--- ormconfig.json.example
@@ -1,16 +1,0 @@
-{
-  "type": "postgres",
-  "host": "localhost",
-  "port": 5432,
-  "username": "postgres",
-  "password": "",
-  "database": "",
-  "logging": true,
-  "entities": ["build/entities/*.js"],
-  "migrations": ["build/migrations/*.js"],
-  "seeds": ["build/seeds/**/*{.ts,.js}"],
-  "factories": ["build/factories/**/*{.ts,.js}"],
-  "cli": {
-    "migrationsDir": "server/migrations"
-  }
-}
