+++ server/configs/index.ts
@@ -1,0 +1,6 @@
+import dotenv from "dotenv";
+import path from "path";
+
+dotenv.config({
+  path: path.join(__dirname, "..", ".env")
+});
